# Conflicts:
#	src/main/java/model/Monde.java
#	src/main/java/model/PieceAttaque.java
#	src/main/java/model/PieceScore.java
#	src/main/java/model/personnages/Pacman.java
#	src/main/java/model/plateau/Case.java
#	src/main/java/model/plateau/Labyrinthe.java

+++ src/main/java/model/plateau/Couloir.java
@@ -15,7 +15,7 @@ public class Couloir extends Case{
     }
 
     @Override
-    public double getCoutAcces() {
+    public int getCoutAcces() {
         return 1;
     }
